pyproject.toml
[tool.poetry]
name = "poke-challenge"
version = "0.1.0"
description = ""
authors = ["Hugo del Solar <hudrec11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
django-environ = "^0.4.5"
psycopg2-binary = "^2.9.1"
requests = "^2.25.1"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
